requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.10.3
pypdf==3.17.4
numpy>=1.24.0
faiss-cpu>=1.12.0
sentence-transformers==2.2.2

